// src/sales/dto/import-sales.dto.ts
export class ImportSaleRow {
    email: string;
    customerName?: string;
    productName?: string;
    sku?: string;
    quantity: string | number;
    unitPrice?: string | number;
    totalValue: string | number;
    paymentMethod?: string;
    status?: string;
    channel?: string;
    date?: string;
}